requirements.txt
boto==2.49.0
boto3==1.15.16
botocore==1.18.16
certifi==2020.6.20
chardet==3.0.4
docutils==0.16
idna==2.10
jenkins==1.0.2
jenkinsapi==0.3.11
jmespath==0.10.0
lxml==4.5.2
pexpect==4.8.0
pkg-resources==0.0.0
ptyprocess==0.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
s3transfer==0.3.3
simplejson==3.17.2
six==1.15.0
urllib3==1.25.10
